move Animation, Font and Texture class to a "accessor" folder
