+++ examples/clientServerLifeSimulator_004/share/includes/C_creature.hpp
@@ -18,9 +18,9 @@
 #define _EXFGE_C_CREATURE_HPP_INCLUDED
 
 #include "C_customObject.hpp"
-#include "FastEngine/C_animation.hpp"
 #include "FastEngine/C_clock.hpp"
-#include "FastEngine/C_font.hpp"
+#include "FastEngine/accessor/C_animation.hpp"
+#include "FastEngine/accessor/C_font.hpp"
 #ifndef FGE_DEF_SERVER
     #include "FastEngine/C_soundBuffer.hpp"
 #endif // FGE_DEF_SERVER
