+++ includes/FastEngine/graphic/C_renderStates.hpp
@@ -17,8 +17,8 @@
 #ifndef _FGE_GRAPHIC_C_RENDERSTATES_HPP_INCLUDED
 #define _FGE_GRAPHIC_C_RENDERSTATES_HPP_INCLUDED
 
-#include "FastEngine/C_texture.hpp"
 #include "FastEngine/C_vector.hpp"
+#include "FastEngine/accessor/C_texture.hpp"
 #include "FastEngine/textureType.hpp"
 #include "FastEngine/vulkan/C_blendMode.hpp"
 #include "FastEngine/vulkan/C_descriptorSet.hpp"
