+++ includes/FastEngine/object/C_objShape.hpp
@@ -26,7 +26,7 @@
 
 #include "FastEngine/fge_extern.hpp"
 #include "C_object.hpp"
-#include "FastEngine/C_texture.hpp"
+#include "FastEngine/accessor/C_texture.hpp"
 
 #define FGE_OBJSHAPE_INSTANCES_SHADER_VERTEX "FGE:OBJ:SHAPE:VERTEX"
 #define FGE_OBJSHAPE_INDEX_FILLCOLOR 0
