+++ includes/FastEngine/object/C_objSprite.hpp
@@ -19,7 +19,7 @@
 
 #include "FastEngine/fge_extern.hpp"
 #include "C_object.hpp"
-#include "FastEngine/C_texture.hpp"
+#include "FastEngine/accessor/C_texture.hpp"
 
 #define FGE_OBJSPRITE_CLASSNAME "FGE:OBJ:SPRITE"
