+++ sources/graphic/C_renderTarget.cpp
@@ -15,7 +15,7 @@
  */
 
 #include "FastEngine/graphic/C_renderTarget.hpp"
-#include "FastEngine/C_texture.hpp"
+#include "FastEngine/accessor/C_texture.hpp"
 #include "FastEngine/extra/extra_function.hpp"
 #include "FastEngine/graphic/C_drawable.hpp"
 #include "FastEngine/graphic/C_transform.hpp"
